japanese_weather_agent/pyproject.toml
[tool.poetry]
name = "japanese-weather-agent"
version = "0.1.0"
description = "日本語の天気情報アシスタント"
authors = ["User <user@example.com>"]
readme = "README.md"
packages = [{include = "japanese_weather_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^0.1.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 